fix(B-301): rename training preparation files to scenario preparation
