+++ backend/app/schemas/scenario_preparation_schema.py
