+++ backend/app/schemas/training_feedback_schema.py
@@ -4,7 +4,7 @@ from app.models.camel_case import CamelModel
 
 
 class ExamplesRequest(CamelModel):
-    transcript: str  # Full transcript of the training session
+    transcript: str  # Full transcript of the session
     objectives: list[str] = Field(
         ..., description='List of training objectives the user is expected to achieve'
     )
@@ -37,7 +37,7 @@ class TrainingExamplesCollection(CamelModel):
 class GoalsAchievementRequest(CamelModel):
     """Request to evaluate the training objectives that were achieved in the transcript."""
 
-    transcript: str  # Full transcript of the training session
+    transcript: str  # Full transcript of the session
     objectives: list[str] = Field(
         ..., description='List of training objectives the user is expected to achieve'
     )
@@ -52,7 +52,7 @@ class GoalsAchievedCollection(CamelModel):
 
 
 class RecommendationsRequest(ExamplesRequest):
-    """Request to generate improvement recommendations based on training feedback.
+    """Request to generate improvement recommendations based on session feedback.
     Same fields as ExamplesRequest, but used for generating recommendations instead of examples.
     """
