+++ backend/app/tests/services/test_scenario_preparation_service_db.py
@@ -7,9 +7,9 @@ from sqlalchemy import create_engine
 from sqlmodel import Session as DBSession
 from sqlmodel import SQLModel
 
-from app.models.training_preparation import ScenarioPreparation, ScenarioPreparationStatus
-from app.schemas.training_preparation_schema import KeyConcept, ScenarioPreparationRequest
-from app.services.training_preparation_service import (
+from app.models.scenario_preparation import ScenarioPreparation, ScenarioPreparationStatus
+from app.schemas.scenario_preparation_schema import KeyConcept, ScenarioPreparationRequest
+from app.services.scenario_preparation_service import (
     create_pending_preparation,
     generate_scenario_preparation,
 )
@@ -54,15 +54,15 @@ class TestScenarioPreparationService(unittest.TestCase):
         self.assertEqual(retrieved.status, ScenarioPreparationStatus.pending)
 
     @patch(
-        'app.services.training_preparation_service.generate_objectives',
+        'app.services.scenario_preparation_service.generate_objectives',
         return_value=['Step 1', 'Step 2'],
     )
     @patch(
-        'app.services.training_preparation_service.generate_checklist',
+        'app.services.scenario_preparation_service.generate_checklist',
         return_value=['Item A', 'Item B'],
     )
     @patch(
-        'app.services.training_preparation_service.generate_key_concept',
+        'app.services.scenario_preparation_service.generate_key_concept',
         return_value=[
             KeyConcept(
                 header='Clear Communication',
@@ -117,7 +117,7 @@ class TestScenarioPreparationService(unittest.TestCase):
         )
 
     @patch(
-        'app.services.training_preparation_service.generate_key_concept',
+        'app.services.scenario_preparation_service.generate_key_concept',
         return_value=[
             KeyConcept(
                 header='Clear Communication',
@@ -134,11 +134,11 @@ class TestScenarioPreparationService(unittest.TestCase):
         ],
     )
     @patch(
-        'app.services.training_preparation_service.generate_checklist',
+        'app.services.scenario_preparation_service.generate_checklist',
         return_value=['Item A', 'Item B'],
     )
     @patch(
-        'app.services.training_preparation_service.generate_objectives',
+        'app.services.scenario_preparation_service.generate_objectives',
         side_effect=RuntimeError('LLM error'),
     )
     def test_generate_scenario_preparation_failed(
